Added an 'env' layer above the block structure symbol table.
This allows creation of arbitrary semantic categories like
:var, :func, :type, or anything EQUAL testable, as a key
in a hash whose value is a block structured symbol table.

Now psilisp can book keep many different kinds of namespaces
as needed during the work on the compiler. For example, if
this were a lisp-2, variables and function namespaces could
be kept apart, etc, etc.