Renamed first psss to alphatization.
alpha-conversion is a renaming of variables in the lambda calculus.
We don't do it as explicitly as that is this code, but we do use a block
structured symbol table to create symbol table entries that effectively
do the same thing. This pass also converts the original source form to
a formal AST and deals with the fact that syntax or primitive ops can
be used as variables.