Checkpoint to getting closure conversion pass in.
Each binding form (LAMBDA, LET, LETREC) and the ast TOPLEVEL form now
contain links to the symbol table at that scope.

Tinkered with when initialization of the toplevel symbol table takes place.

Small bits of other kinds of refactoring.

Lots of notes and example codes to concretize how closures work in the
compiler.

Added tentative closure representation in the AST IR, but it isn't complete
yet. Started to add the closure conversion pass.