Add an ability to add syments to :outer-scope. Fixup undefined vars.
Now when we see an undefined variable reference, we assume it is a
global variable (both in syment description and inserting it into the
outer most scope) to prevent the closure algorithm from going wrong or
doing other bad behaviors. It is a plausible understanding that all
undefined vars are ultimately global.