Completed the core of the closure conversion algorithm.
The free variables in the lambdas are now closed over and the
copy propogation of the free variables and their closed versions
are shoved into the symbol table. After this pass, all lambda
functions semantically change to being "closed".

Must finish specifying and complete transmogrification to CLOSURE object.
Notably, this includes a real AST representation of the FVCL info
in each lambda-syntax node in the closure node too.