Formalized the rest of the closure information.
The closure object now contains a specific closure env var that is placed
into the symtab of the function. This var has the id number that the
participating closed vars reference. All vlosed vars and the single closed
env that share the same closure-id are grouped into a closure.

Later passes deal with the actual closure representation and whether
or not closed variables could be made unclosed and passed as arguments.
Basically this initial pass is the normal form of the closure.